banco/requirements.txt
Django>=5.0,<7
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
